+++ src/lib/models.ts
@@ -1,4 +1,5 @@
 export interface WebviewMetadata {
     id: string;
+    title: string;
     src: string;
 }
